requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Cython==0.29.28
Flask==2.2.2
gensim==4.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.1
MarkupSafe==2.1.1
nltk==3.7
numpy==1.23.5
pandas==1.5.2
PyPDF2==2.11.2
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
smart-open==6.2.0
threadpoolctl==3.1.0
tqdm==4.64.1
underthesea==1.3.5
underthesea_core==0.0.5a2
Unidecode==1.3.6
urllib3==1.26.13
Werkzeug==2.2.2
